Creating StartAlreadyInspectionedInspection style file
