+++ src/Styles/StartInspection/StartAlreadyInspectionedInspectionStyles.js
@@ -1,0 +1,70 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+    principal: {
+      flex: 1,
+    },
+  
+    content: {
+      flex: 1,
+      paddingTop: 6,
+      backgroundColor: 'white',
+    },
+  
+    listSchedule: {
+      flex: 1,
+      marginHorizontal: 15,
+      marginVertical: 10,
+      borderColor: 'black',
+      borderWidth: 1,
+      borderRadius: 3,
+      backgroundColor: '#FAFAFA',
+      justifyContent: 'space-between',
+      flexDirection: 'row',
+      alignItems: 'center',
+    },
+  
+    popUp: {
+      marginBottom: 120,
+    },
+  
+    footerPopUp: {
+      backgroundColor: '#F9F9FB',
+      borderColor: '#DAD9DC',
+      borderTopWidth: 0.5,
+      borderBottomLeftRadius: 8,
+      borderBottomRightRadius: 8,
+      flexDirection: 'row',
+      alignItems: 'center',
+      justifyContent: 'center',
+    },
+  
+    textBox: {
+      flex: 6,
+      paddingLeft: 4,
+      justifyContent: 'flex-start',
+      marginRight: 18,
+    },
+  
+    text: {
+      fontSize: 15,
+      paddingVertical: 2,
+    },
+  
+    buttonBox: {
+      borderColor: 'black',
+      borderWidth: 1,
+      borderRadius: 7,
+      backgroundColor: '#DEDEDE',
+      padding: 8,
+      justifyContent: 'center',
+      marginRight: 10,
+    },
+  
+    buttonText: {
+      fontSize: 12,
+      textAlign: 'center',
+    },
+});
+
+export default styles;
