+++ src/screens/startInspection/StartAlreadyInspectionedInspection.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import {
-  StyleSheet,
   Text,
   View,
   TouchableOpacity,
@@ -15,73 +14,7 @@ import PropTypes from 'prop-types';
 import stylesList from '../../Styles/ListStyles';
 import PDFReport from '../../components/PDFReport';
 import LoadingIndicator from '../../components/LoadingIndicator';
-
-const styles = StyleSheet.create({
-  principal: {
-    flex: 1,
-  },
-
-  content: {
-    flex: 1,
-    paddingTop: 6,
-    backgroundColor: 'white',
-  },
-
-  listSchedule: {
-    flex: 1,
-    marginHorizontal: 15,
-    marginVertical: 10,
-    borderColor: 'black',
-    borderWidth: 1,
-    borderRadius: 3,
-    backgroundColor: '#FAFAFA',
-    justifyContent: 'space-between',
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-
-  popUp: {
-    marginBottom: 120,
-  },
-
-  footerPopUp: {
-    backgroundColor: '#F9F9FB',
-    borderColor: '#DAD9DC',
-    borderTopWidth: 0.5,
-    borderBottomLeftRadius: 8,
-    borderBottomRightRadius: 8,
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-
-  textBox: {
-    flex: 6,
-    paddingLeft: 4,
-    justifyContent: 'flex-start',
-    marginRight: 18,
-  },
-
-  text: {
-    fontSize: 15,
-    paddingVertical: 2,
-  },
-
-  buttonBox: {
-    borderColor: 'black',
-    borderWidth: 1,
-    borderRadius: 7,
-    backgroundColor: '#DEDEDE',
-    padding: 8,
-    justifyContent: 'center',
-    marginRight: 10,
-  },
-
-  buttonText: {
-    fontSize: 12,
-    textAlign: 'center',
-  },
-});
+import styles from '../../Styles/StartInspection/StartAlreadyInspectionedInspectionStyles'
 
 class StartAlreadyInspectionedInspection extends React.Component {
   constructor(props) {
